* src/runtime/alloc.h: replaced longs with word-sized integers (sword_t, uword_t)
* src/runtime/backtrace.c (stack_pointer_p, ra_pointer_p): ditto
  (debug_function_from_pc): ditto
  (backtrace_from_fp): adjusted for LP64

* src/runtime/breakpoint.c (compute_offset): ditto

* src/runtime/core.h (ndir_entry): ditto

+++ src/runtime/core.h
@@ -17,11 +17,11 @@
 
 struct ndir_entry {
 #ifndef LISP_FEATURE_ALPHA
-    long identifier;
-    long nwords;
-    long data_page;
-    long address;
-    long page_count;
+    intptr_t identifier;
+    intptr_t nwords;
+    intptr_t data_page;
+    intptr_t address;
+    intptr_t page_count;
 #else
     u32 identifier;
     u32 nwords;
